requirements.txt
biopython
argparse
csv
